+++ controlplane/src/core/repositories/SubgraphRepository.ts
@@ -700,6 +700,7 @@ export class SubgraphRepository {
         hasClientTraffic: true,
         forcedSuccess: true,
         ghDetails: true,
+        hasLintErrors: true,
       },
       limit,
       offset,
@@ -735,6 +736,7 @@ export class SubgraphRepository {
               checkRunId: c.ghDetails.checkRunId,
             }
           : undefined,
+        hasLintErrors: c.hasLintErrors ?? false,
       })),
       checksCount,
     };
@@ -834,6 +836,7 @@ export class SubgraphRepository {
             checkRunId: check.ghDetails.checkRunId,
           }
         : undefined,
+      hasLintErrors: check.hasLintErrors ?? false,
     };
   }