+++ router-tests/testenv/testenv.go
@@ -77,6 +77,7 @@ type Config struct {
 	OverrideAbsinthePath               string
 	ModifyRouterConfig                 func(routerConfig *nodev1.RouterConfig)
 	ModifyEngineExecutionConfiguration func(engineExecutionConfiguration *config.EngineExecutionConfiguration)
+	ModifySecurityConfiguration        func(securityConfiguration *config.SecurityConfiguration)
 	ModifyCDNConfig                    func(cdnConfig *config.CDNConfiguration)
 	DisableWebSockets                  bool
 	TLSConfig                          *core.TlsConfig
@@ -439,6 +440,10 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 		testConfig.ModifyEngineExecutionConfiguration(&engineExecutionConfig)
 	}
 
+	if testConfig.ModifySecurityConfiguration != nil {
+		testConfig.ModifySecurityConfiguration(&cfg.SecurityConfiguration)
+	}
+
 	routerOpts := []core.Option{
 		core.WithStaticRouterConfig(routerConfig),
 		core.WithLogger(zapLogger),
@@ -446,12 +451,13 @@ func configureRouter(listenerAddr string, testConfig *Config, routerConfig *node
 		core.WithDevelopmentMode(true),
 		core.WithPlayground(true),
 		core.WithEngineExecutionConfig(engineExecutionConfig),
+		core.WithSecurityConfig(cfg.SecurityConfiguration),
 		core.WithCDN(cfg.CDN),
 		core.WithListenerAddr(listenerAddr),
 		core.WithTLSConfig(testConfig.TLSConfig),
 		core.WithEvents(config.EventsConfiguration{
-			Sources: []config.EventSource{
-				{
+			Sources: map[string]config.EventSource{
+				"my-nats": {
 					Provider: "NATS",
 					URL:      nats.ClientURL(),
 				},